Removed the Guests Fetchher that's not used.
